requirements.txt
pyautogui
pillow==9.5.0
openai
python-dotenv